1. Terrain shaders added. 2. Binary resource files added. 3. Bullet physics ported.

+++ src/nativeMain/kotlin/org/sadgames/engine/render/gl/fbo/DepthBufferFBO.kt
@@ -7,6 +7,6 @@ import org.sadgames.engine.render.gl.material.textures.DepthTexture
 import org.sadgames.engine.utils.Color4f
 
 open class DepthBufferFBO(width: Int, height: Int, color: Color4f = DEPTH_BUFFER_CLEAR_COLOR, isMultiSampled: Boolean = false): AbstractFBO(width, height, color, isMultiSampled = isMultiSampled) {
-    override fun attachTexture(num: Int) = DepthTexture(width, height, isMultiSampled)(FBO_TEXTURE_SLOT, true)
+    override fun attachTexture(num: Int) = DepthTexture(width, height, isMultiSampled)(FBO_TEXTURE_SLOT.toInt(), true)
     override fun getBltMask() = GL_DEPTH_BUFFER_BIT
 }
